+++ popup.js
@@ -20,7 +20,7 @@ document.addEventListener('DOMContentLoaded', function () {
                                 url: "https://www.baseball-reference.com"
                             });
                         } else {
-                            var lastName = player.name_last;
+                            var lastName = player.name_last.toLowerCase();;
                             var initial = lastName[0];
 
                             chrome.tabs.update({
@@ -46,7 +46,30 @@ document.addEventListener('DOMContentLoaded', function () {
                                 url: "https://www.baseball-reference.com"
                             });
                         } else {
-                            var lastName = player.name_last;
+                            var lastName = player.name_last.toLowerCase();;
+                            var initial = lastName[0];
+
+                            chrome.tabs.update({
+                                url: `https://www.baseball-reference.com/players/${initial}/${player.key_bbref}.shtml`
+                            });
+                        }
+                    }
+                    );
+            }
+            else if (currentLocation.substring(12, 30) === "baseballprospectus") {
+                var id = getBaseballProspectusId(currentLocation, id);
+
+                fetch("../data/bpid.json")
+                    .then(response => response.json())
+                    .then(json => {
+                        var player = json[id];
+
+                        if (player === undefined) {
+                            chrome.tabs.update({
+                                url: "https://www.baseball-reference.com"
+                            });
+                        } else {
+                            var lastName = player.name_last.toLowerCase();
                             var initial = lastName[0];
 
                             chrome.tabs.update({
@@ -84,7 +107,7 @@ document.addEventListener('DOMContentLoaded', function () {
                                 url: "https://www.baseball-reference.com"
                             });
                         } else {
-                            var lastName = player.name_last;
+                            var lastName = player.name_last.toLowerCase();;
                             var initial = lastName[0];
 
                             chrome.tabs.create({
@@ -110,7 +133,30 @@ document.addEventListener('DOMContentLoaded', function () {
                                 url: "https://www.baseball-reference.com"
                             });
                         } else {
-                            var lastName = player.name_last;
+                            var lastName = player.name_last.toLowerCase();;
+                            var initial = lastName[0];
+
+                            chrome.tabs.create({
+                                url: `https://www.baseball-reference.com/players/${initial}/${player.key_bbref}.shtml`
+                            });
+                        }
+                    }
+                    );
+            }
+            else if (currentLocation.substring(12, 30) === "baseballprospectus") {
+                var id = getBaseballProspectusId(currentLocation, id);
+
+                fetch("../data/bpid.json")
+                    .then(response => response.json())
+                    .then(json => {
+                        var player = json[id];
+
+                        if (player === undefined) {
+                            chrome.tabs.create({
+                                url: "https://www.baseball-reference.com"
+                            });
+                        } else {
+                            var lastName = player.name_last.toLowerCase();;
                             var initial = lastName[0];
 
                             chrome.tabs.create({
@@ -129,7 +175,6 @@ document.addEventListener('DOMContentLoaded', function () {
         chrome.tabs.query(query, callback);
         return false;
     }, false);
-
 
     var fanButton = document.getElementById('savant');
     fanButton.addEventListener('click', function () {
@@ -181,6 +226,26 @@ document.addEventListener('DOMContentLoaded', function () {
             }
             else if (currentLocation.substring(8, 22) === "baseballsavant") {
                 //They clicked the savant button, so they're already here?
+            }
+            else if (currentLocation.substring(12, 30) === "baseballprospectus") {
+                var id = getBaseballProspectusId(currentLocation, id);
+
+                fetch("../data/bpid.json")
+                    .then(response => response.json())
+                    .then(json => {
+                        var player = json[id];
+
+                        if (player === undefined) {
+                            chrome.tabs.update({
+                                url: "https://baseballsavant.mlb.com"
+                            });
+                        } else {
+                            chrome.tabs.update({
+                                url: `https://baseballsavant.mlb.com/savant-player/${player.name_first}-${player.name_last}-${player.key_mlbam}`
+                            });
+                        }
+                    }
+                    );
             }
             else {
                 chrome.tabs.update({
@@ -239,6 +304,26 @@ document.addEventListener('DOMContentLoaded', function () {
             }
             else if (currentLocation.substring(8, 22) === "baseballsavant") {
                 //They clicked the savant button, so they're already here?
+            }
+            else if (currentLocation.substring(12, 30) === "baseballprospectus") {
+                var id = getBaseballProspectusId(currentLocation, id);
+
+                fetch("../data/bpid.json")
+                    .then(response => response.json())
+                    .then(json => {
+                        var player = json[id];
+
+                        if (player === undefined) {
+                            chrome.tabs.create({
+                                url: "https://baseballsavant.mlb.com"
+                            });
+                        } else {
+                            chrome.tabs.create({
+                                url: `https://baseballsavant.mlb.com/savant-player/${player.name_first}-${player.name_last}-${player.key_mlbam}`
+                            });
+                        }
+                    }
+                    );
             }
             else {
                 chrome.tabs.create({
@@ -301,6 +386,26 @@ document.addEventListener('DOMContentLoaded', function () {
                     }
                     );
             }
+            else if (currentLocation.substring(12, 30) === "baseballprospectus") {
+                var id = getBaseballProspectusId(currentLocation, id);
+
+                fetch("../data/bpid.json")
+                    .then(response => response.json())
+                    .then(json => {
+                        var player = json[id];
+
+                        if (player === undefined) {
+                            chrome.tabs.update({
+                                url: "https://www.fangraphs.com"
+                            });
+                        } else {
+                            chrome.tabs.update({
+                                url: `https://www.fangraphs.com/players/${player.name_first}-${player.name_last}/${player.key_fangraphs}/stats`
+                            });
+                        }
+                    }
+                    );
+            }
             else {
                 chrome.tabs.update({
                     url: "https://www.fangraphs.com"
@@ -359,9 +464,187 @@ document.addEventListener('DOMContentLoaded', function () {
                     }
                     );
             }
+            else if (currentLocation.substring(12, 30) === "baseballprospectus") {
+                var id = getBaseballProspectusId(currentLocation, id);
+
+                fetch("../data/bpid.json")
+                    .then(response => response.json())
+                    .then(json => {
+                        var player = json[id];
+
+                        if (player === undefined) {
+                            chrome.tabs.create({
+                                url: "https://www.fangraphs.com"
+                            });
+                        } else {
+                            chrome.tabs.create({
+                                url: `https://www.fangraphs.com/players/${player.name_first}-${player.name_last}/${player.key_fangraphs}/stats`
+                            });
+                        }
+                    }
+                    );
+            }
             else {
                 chrome.tabs.create({
                     url: "https://www.fangraphs.com"
+                });
+            }
+        }
+        chrome.tabs.query(query, callback);
+        return false;
+    }, false);
+
+    var fanButton = document.getElementById('bp');
+    fanButton.addEventListener('click', function () {
+        var query = { active: true, currentWindow: true };
+        function callback(tabs) {
+            var currentTab = tabs[0];
+            console.log(currentTab.url);
+            var currentLocation = currentTab.url;
+            if (currentLocation.substring(12, 30) === "baseballprospectus") {
+                //They clicked the baseballprospectus button, so they're already here?
+            }
+            else if (currentLocation.substring(12, 30) === "baseball-reference") {
+                var id = getBBRefId(currentLocation, id);
+
+                fetch("../data/bbref.json")
+                    .then(response => response.json())
+                    .then(json => {
+                        var player = json[id];
+
+                        if (player === undefined) {
+                            chrome.tabs.update({
+                                url: "https://www.baseballprospectus.com"
+                            });
+                        } else {
+                            chrome.tabs.update({
+                                url: `https://www.baseballprospectus.com/player/${player.bpid}`
+                            });
+                        }
+                    }
+                    );
+            }
+            else if (currentLocation.substring(8, 22) === "baseballsavant") {
+                var id = getSavantId(currentLocation, id);
+
+                fetch("../data/mlbam.json")
+                    .then(response => response.json())
+                    .then(json => {
+                        var player = json[id];
+
+                        if (player === undefined) {
+                            chrome.tabs.update({
+                                url: "https://www.baseballprospectus.com"
+                            });
+                        } else {
+                            chrome.tabs.update({
+                                url: `https://www.baseballprospectus.com/player/${player.bpid}`
+                            });
+                        }
+                    }
+                    );
+            }
+            else if (currentLocation.substring(12, 21) === "fangraphs") {
+                var id = getFangraphsId(currentLocation, id);
+
+                fetch("../data/fangraphs.json")
+                    .then(response => response.json())
+                    .then(json => {
+                        var player = json[id];
+
+                        if (player === undefined) {
+                            chrome.tabs.update({
+                                url: "https://www.baseballprospectus.com"
+                            });
+                        } else {
+                            chrome.tabs.update({
+                                url: `https://www.baseballprospectus.com/player/${player.bpid}`
+                            });
+                        }
+                    }
+                    );
+            }
+            else {
+                chrome.tabs.update({
+                    url: "https://www.baseballprospectus.com"
+                });
+            }
+        }
+        chrome.tabs.query(query, callback);
+    }, false);
+    fanButton.addEventListener('contextmenu', function (ev) {
+        ev.preventDefault();
+        var query = { active: true, currentWindow: true };
+        function callback(tabs) {
+            var currentTab = tabs[0];
+            console.log(currentTab.url);
+            var currentLocation = currentTab.url;
+            if (currentLocation.substring(12, 30) === "baseballprospectus") {
+                //They clicked the baseballprospectus button, so they're already here?
+            }
+            else if (currentLocation.substring(12, 30) === "baseball-reference") {
+                var id = getBBRefId(currentLocation, id);
+
+                fetch("../data/bbref.json")
+                    .then(response => response.json())
+                    .then(json => {
+                        var player = json[id];
+
+                        if (player === undefined) {
+                            chrome.tabs.create({
+                                url: "https://www.baseballprospectus.com"
+                            });
+                        } else {
+                            chrome.tabs.create({
+                                url: `https://www.baseballprospectus.com/player/${player.bpid}`
+                            });
+                        }
+                    }
+                    );
+            }
+            else if (currentLocation.substring(8, 22) === "baseballsavant") {
+                var id = getSavantId(currentLocation, id);
+
+                fetch("../data/mlbam.json")
+                    .then(response => response.json())
+                    .then(json => {
+                        var player = json[id];
+
+                        if (player === undefined) {
+                            chrome.tabs.create({
+                                url: "https://www.baseballprospectus.com"
+                            });
+                        } else {
+                            chrome.tabs.create({
+                                url: `https://www.baseballprospectus.com/player/${player.bpid}`
+                            });
+                        }
+                    }
+                    );
+            }
+            else if (currentLocation.substring(12, 21) === "fangraphs") {
+                var id = getFangraphsId(currentLocation, id);
+
+                fetch("../data/fangraphs.json")
+                    .then(response => response.json())
+                    .then(json => {
+                        var player = json[id];
+
+                        if (player === undefined) {
+                            chrome.tabs.create({
+                                url: "https://www.baseballprospectus.com"
+                            });
+                        } else {
+                            chrome.tabs.create({
+                                url: `https://www.baseballprospectus.com/player/${player.bpid}`
+                            });
+                        }
+                    }
+                    );
+            }
+            else {
+                chrome.tabs.create({
+                    url: "https://www.baseballprospectus.com"
                 });
             }
         }
@@ -407,5 +690,15 @@ function getFangraphsId(currentLocation) {
     console.log(indexOfSecondSlash);
     var id = nameAndId.substring(0, indexOfSecondSlash);
     console.log(id);
+    return id;
+}
+
+function getBaseballProspectusId(currentLocation) {
+    var locationOfPlayerString = currentLocation.indexOf('player');
+    var playerStringTillEnd = currentLocation.substring(locationOfPlayerString + 7);
+    console.log(playerStringTillEnd);
+    var indexOfSlash = playerStringTillEnd.indexOf("/");
+    console.log(indexOfSlash);
+    var id = playerStringTillEnd.substring(0, indexOfSlash);
     return id;
 }